pr reccomandations implemented
